Convert landing page to zero-JS Next.js application.
