+++ packages/bun-landing/README.md
@@ -1,0 +1,25 @@
+# bun.sh
+
+This is the landing page for the bun.sh site and documentation.
+
+## Running Locally
+
+Install dependencies
+
+```bash
+bun install
+```
+
+Run the development server:
+
+```bash
+bun dev
+```
+
+Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+
+## Building
+
+```bash
+bun run build
+```
