+++ packages/bun-landing/bunfig.toml
@@ -1,0 +1,1 @@
+framework = "next"
