+++ packages/bun-landing/pages/_app.tsx
@@ -1,0 +1,6 @@
+import Layout from "../components/Layout";
+import "../styles/global.css";
+
+export default function MyApp({ Component, pageProps }) {
+  return <Layout><Component {...pageProps} /></Layout>;
+}
